+++ src/relay/src/compiler/createPrintRequireDefaultModuleDependency.js
@@ -1,0 +1,5 @@
+// @flow strict
+
+export default function createPrintRequireDefaultModuleDependency(): (string) => string {
+  return (moduleName) => `require('./${moduleName}.js').default`;
+}