+++ src/relay/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Unreleased
 
+- removed required `X-Client` HTTP header from `createNetworkFetcher` (you can still use it but it's no longer mandatory)
+- relax signature verification when using `adeira-relay-compiler` - it no longer requires the signature but still verifies it if it exists
+
+# 3.1.1
+
+- fixes a Flow issue when using `@refetchable` (`useRefetchableFragment` hook), see: https://github.com/adeira/universe/issues/1975
+
 # 3.1.0
 
 This release introduces Relay Hooks in a backward compatible fashion. Next major version of `@adeira/relay` will focus on phasing out legacy parts of Relay Modern (containers API), so it's highly encouraged to upgrade as soon as possible.